# api-specification.yaml
openapi: 3.0.0
info:
  title: Distributed Assessment Platform API
  version: 1.0.0
  description: |
    Scalable distributed assessment platform with automated notifications, reporting, and data export capabilities.
    
    ## Features
    - Role-based access control (Admin/Student)
    - Hierarchical content organization (Subject→Chapter→Quiz→Question)
    - Real-time quiz attempts with automatic scoring
    - Automated background jobs for daily reminders and monthly reports
    - Data export capabilities with CSV generation
    
    ## Authentication
    All endpoints require JWT Bearer token authentication except for login and registration.
    
  contact:
    name: Distributed Assessment Platform Support
    email: support@assessmentplatform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.assessmentplatform.com
    description: Production server
...
paths:
...
  /admin/quiz/{quiz_id}/question:
    get:
      tags: [Admin - Quiz]
      summary: Get all questions for a specific quiz
      parameters:
        - name: quiz_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of questions for quiz
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionListResponse'
...
  /admin/question:
    get:
      tags: [Admin - Question]
      summary: Get all questions
      responses:
        '200':
          description: List of questions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionListResponse'
    post:
      tags: [Admin - Question]
      summary: Create a new question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionRequest'
            example:
              quiz_id: 1
              question: What is 2+2?
              option1: "3"
              option2: "4"
              option3: "5"
              option4: "6"
              answer: 2
              marks: 1
      responses:
        '201':
          description: Question created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
  /admin/question/{question_id}:
    get:
      tags: [Admin - Question]
      summary: Get question by ID
      parameters:
        - name: question_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Question details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
    put:
      tags: [Admin - Question]
      summary: Update question
      parameters:
        - name: question_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionRequest'
      responses:
        '200':
          description: Question updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
    delete:
      tags: [Admin - Question]
      summary: Delete question
      parameters:
        - name: question_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Question deleted successfully
  /admin/question/search:
    get:
      tags: [Admin - Question]
      summary: Search questions
      parameters:
        - name: text
          in: query
          required: true
          schema:
            type: string
            minLength: 3
      responses:
        '200':
          description: Question search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionListResponse'

  # Admin - Dashboard
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    QuestionRequest:
      type: object
      required:
        - quiz_id
        - question
        - option1
        - option2
        - option3
        - option4
        - answer
        - marks
      properties:
        quiz_id:
          type: integer
        question:
          type: string
          maxLength: 4096
        option1:
          type: string
          maxLength: 4096
        option2:
          type: string
          maxLength: 4096
        option3:
          type: string
          maxLength: 4096
        option4:
          type: string
          maxLength: 4096
        answer:
          type: integer
          minimum: 1
          maximum: 4
          description: Correct answer option number (1-4)
        marks:
          type: integer
          minimum: 1
          description: Points awarded for correct answer
    QuestionResponse:
      type: object
      properties:
        id:
          type: integer
        quiz_id:
          type: integer
        quiz_name:
          type: string
        chapter_id:
          type: integer
        chapter_name:
          type: string
        subject_id:
          type: integer
        subject_name:
          type: string
        question:
          type: string
        option1:
          type: string
        option2:
          type: string
        option3:
          type: string
        option4:
          type: string
        answer:
          type: integer
        marks:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    QuestionListResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/QuestionResponse'

    # Authentication Schemas
...
security:
  - BearerAuth: []